Attach comments to elements of a MojomFile.
The primary purpose of this change is AttachComments function. It accepts a MojomFile and a list of comment tokens and uses some "reasonable" heuristics to attach the comments to the right element in the mojom file. See the comment on the function for details.

The AttachedCommentsHolder interface is implemented by all mojom file elements. Long story short, it allows one to attach comments above, to the right or to the left of an element.

NewAttributes is updated to take the left square bracket token that starts the attribute section in the mojom file. The allows us to know when attributes is started.

BUG= #592
R=rudominer@chromium.org

Review URL: https://codereview.chromium.org/1664833003 .