mojom_types.mojom: Changes the name |interface_name| to |service_name| in struct MojomInterface.
Also the mojom parser now populates this field with the correct service name (that is, the value of the "ServiceName" attribute) if it is set. The field was changed to nullable and will be unset if the interface does not have the attribute.

We removed the logic from mojom_translator.py that gets the service name from the "ServiceName" attribute and instead get it from the service_name field (which was set to the value in the "ServiceName" attribute in the parser.) The
interface_name field of mojom.Interface is now obtained from the decl_data.short_name field. Previously it was obtained from the interface_name field (which was set to the value of decl_data.short_name in the parser.)

BUG=#448
R=azani@chromium.org

Review URL: https://codereview.chromium.org/1677343002 .