Refactor the view manager interfaces.
This change brings the interfaces more in line with the conventions
adopted by the compositor.  In particular, it deemphasizes the role
of the old |View| interface by renaming it to |ViewListener|,
meanwhile |ViewHost| becomes |View|.  Analoguous changes are made
to |ViewTree| and |ViewTreeHost|.

The reason to do this is that the listeners may eventually be split
up or become optional depending on how the views or view trees are
configured so it doesn't make sense to think of them as monolithic
components anymore.

Also significantly improved the documentation of the mojom files.

BUG=
R=abarth@google.com

Review URL: https://codereview.chromium.org/1702133004 .