Use the new mojom parser from check_mojom_dart.py.
BUG=#671

This is part of my work to delete the old Mojom parser from the source tree. Currently the pre-submit script mojo/dart/tools/presubmit/check_mojom_dart.py directly invokes v1 of the parser in order to extract the module name and the 'DartPackage' attribute of a .mojom file.

This CL changes check_mojom_dart.py to use the new parser. In support of that we also:
- Create a new script in pylib called parser_runner.py that encapsulates invoking the parser and the translator.
- Invoke this script from check_mojom_dart.py
- Refactor run_code_generators.py and mojom_bindings_generator_v2.py to invoke parser_runner.py instead of having the logic directly embbedded in them.

R=azani@chromium.org, johnmccutchan@google.com

Review URL: https://codereview.chromium.org/1700443003 .