Ensure presentation time is monotonic.
Previously the frame scheduler explicitly disclaimed responsibility
for ensuring that the presentation time is monotonic in cases where
the pipeline delays have changed (which should be rare).

However, such a policy is brittle and error prone for the same reason
that daylight saving time causes such headaches in the fall.  So instead
promise to be monotonic and try to do something sensible to catch up.

It's likely that this policy will require further refinement in the
future, especially if large changes in pipeline latency become
commonplace.  We need to keep in mind that the presentation time stamp
mechanism is also intended to be used to achieve A/V synchronization so
the timing model must be accurate and reliable.

BUG=
R=abarth@google.com

Review URL: https://codereview.chromium.org/1714823004 .