A few fixes to the mojom formatter.
Update the way multiline comments are aligned. In the process fix the issue
where space are added on empty lines in multiline comments.

Fix the build tool to reflect the new location of the sha1 files.

Fix the printer to print nothing if a module's namespace is the empty string.

R=rudominer@chromium.org, rudominer
BUG= #592

Review URL: https://codereview.chromium.org/1728853007 .