Change Mojo main.
The first argument to main should be a List<String>. As such the mojo handle doesn't fit there.

Dart main functions may have an additional message-argument which is perfectly suited for the mojo-handle.
The Mojo script isn't necessary for the main either, since Platform.script can return that one. This makes the 'args' argument contain the arguments that were passed to the Mojo application via the command-line (which is what is expected).

R=zra@google.com

Review URL: https://codereview.chromium.org/1705033003 .