Go bindings: Add accessors for the UserDefinedType to a Struct, Enum, Interface and Union
This change is to support use cases such as SyncBase in which we have the Go representation of a Mojo object and we want to get the UserDefinedType of its Mojom type to facilitate serializing it to VOM.

We add both static and non-static accessors. We add non-static accessors directly to the Go type for structs
and enums. For unions we add the non-static accessors to each of the variant types since the union type itself
is an interface. For Mojom interfaces we add the methods to the interface request type.

We do not expose a method for obtaining a type key in order to discourage passing type keys around (since a type key is only guaranteed to have meaning in a limited context). Instead we directly return the userDefinedType.

This also allows us to get rid of the hack in validation_type_test.go where type keys are being manufactured.

- We add a new template file called runtime_type_macros.tmpl and include it in enum.tmpl, interface.tml, union.tmpl and struct.tmpl
- Remove computeTypeKey() from validation_type_test.go and instead use the static accessor functions.
- Add additional test for the non-static accessors.

BUG=#448
R=alexfandrianto@google.com, azani@chromium.org

Review URL: https://codereview.chromium.org/1733243004 .