Make tools/run_hooks.py work when depot_tools is not in PATH
It should be possible to create and build from a Mojo checkout even
when depot_tools is not in the PATH, even though many common
development workflows require it. This teaches tools/run_hooks.py to
add a default depot_tools location to the PATH that hooks run in so
that a jiri checkout can execute hooks without PATH modifications.

R=kulakowski@chromium.org, lanechr@google.com, viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/1760353002 .