Added ImagePipe
ImagePipe is a Mojo interface designed to provide synchronized transport of
graphics buffers residing on the GPU (though in principle it could be used for
the transport of any type media buffer that can be accessed via a mojo handle).
The semantics of this interface are designed mainly to support the creation of
accelerated rendering contexts (e.g. OpenGL contexts) whose output is consumed
by a different Mojo application than the one doing the rendering.

This interface can be used directly, but we also provide endpoint classes which
provide state tracking and state validation functionality, which should
generally be used unless there is a compeeling reason not to.

Change-Id: I389ec3519314da2623a65051ef9786282a28e3ac

BUG=
R=jeffbrown@google.com

Review URL: https://codereview.chromium.org/1595773002 .