Go bindings: Adds support for optional deterministic encoding.
Mojo message serialization is non-deterministic (at least in Go) because the order of the keys and values in a map is not specified. For the sake of IPC this doesn't matter. But for some other use cases it is better for it to be deterministic.

The use case I have currently is that in the implementation of Mojom runtime type info we write a literal into the generated source files that encodes the bytes of a serialized Mojo struct containing the type info. If these bytes are not deterministic it means that the output of the code generator is not stable (i.e. it may change even if the input .mojom file has not changed.) This causes problems for situations such as the generation of the .mojom.dart files that are checked in to source control.

In this CL we add a SetDeterministic() method to Encoder(). If SetDeterministic(true) is invoked then when a Mojo map is serialized into an array of keys and an array of values, the arrays are sorted in ascending key order. This will not affect normal Mojo IPC at all because the author of a Mojo app in general never needs to directly interact with an Encoder.

R=azani@chromium.org, bprosnitz@google.com

Review URL: https://codereview.chromium.org/1778813002 .