Mozart: Directly associate the renderer with the view tree.
Instead of hooking up the view tree to the renderer indirectly
in the launcher application, give the Renderer interface to the
View Manager to take care of instead.

This change will allow the view manager to perform hit testing
on the scene graph.  It also simplifies the launcher somewhat.

BUG=
R=abarth@google.com

Review URL: https://codereview.chromium.org/1775143004 .