Mozart: Make Snapshot immutable.
Follow the same pattern as the rest of the scene graph content
objects.  This will make it easier to implement hit testing safely.

Moved the RenderFrame out of the Snapshot since it can go directly
to the renderer then we don't need it anymore afterwards.

BUG=
R=abarth@google.com

Review URL: https://codereview.chromium.org/1778793002 .