Mozart: Specify texture origin explicitly.
Previously we assumed the origin was always top-left but OpenGL
likes to create textures with the origin in the bottom-left.
Since we don't always get what we want in this life, let the client
specify what origin it used.  Maybe later we can grumble about it
and replace (or augment) this with a more general means of specifying
texture coordinates or texture matrices for rendering.

Flipping the node's coordinate system upside-down in Spinning Cube
has undesirable consequences on the input dispatch (since the events
get flipped too, as expected) so it was time to remove the old hack
and resolve the origin problem.

BUG=
R=abarth@google.com

Review URL: https://codereview.chromium.org/1776643004 .