Mojom parser: Compute and validate union field tags.
In mojom_types.mojom union fields are supposed to be listed in tag order, and the
original lexical order is preserved in a field of DeclarationData called |declaration_order|.
Also union fields have a field called |Tag|.

Additionally there is a field of DeclarationData called |declared_ordinal| that records if an
explicit ordinal has been specified.

Previously only the |declared_ordinal| bit was implemented. This has worked so far because the
actual field tags are currently computed in the Python back-end using the |declared_ordinal| field.

This CL implements the missing bits which will allow us in a follow-up CL to eliminate the logic from the Python backend. Implementing the missing bits is also important because we are starting to re-write the backends in Go
and also because the runtime type information requires it.

This CL is part two of this project. Part 1 was https://codereview.chromium.org/1767033002/.

BUG=#696
R=azani@chromium.org

Review URL: https://codereview.chromium.org/1780913004 .