Mojom compiler: Stop re-computing union tags in the backend.
Union tag values are being computed in the frontend now. This patch has us stop re-computing the tags in the backend and start simply consuming the values from the MojomFileGraph.

We didn't change the templates so the tag value is still in an attribute named |ordinal|.

NOTE: Users of the mojom parser must invoke "gclient sync" after pulling this change in order to obtain the latest
version of the frontend. The changes to the frontend were in https://codereview.chromium.org/1780913004/.
BUG=#713
R=azani@chromium.org

Review URL: https://codereview.chromium.org/1803663003 .