Mojom compiler: Add additional GN dependencies in the "mojom" target.
The Mojom generator needs to be re-run if the frontend of the compiler changes. Also it needs to be re-run if mojom_types.mojom changes and generate_type_info=true.

At the moment the Mojo waterfall is broken and I am trying to fix it. I am unable to reproduce the test failure locally and I'm hypothesizing that the problem is that some .mojom.go files need to be re-generated on the bots. I hope this will trigger that.

TBR=azani@chromium.org

bug=#715

Review URL: https://codereview.chromium.org/1804833002 .