Teach //examples/shadows to draw a green square
We must first draw a green square before we can draw a green square that casts
a shadow.

This code in this CL depends on glm, which has not yet been added to DEPS.
Adding glm to DEPS is blocked on getting a mirror set up, which is in progress.

R=kulakowski@chromium.org

Review URL: https://codereview.chromium.org/1805803003 .