Check that the mojom tool binary is up to date before running.
If mojom.sha1 can be found in the same directory as the mojom tool binary,
the mojom tool checks that its own sha1 sum matches that in mojom.sha1. If that
is not the case, the mojom tool exits and prints an error message.

If there is no mojom.sha1 file in the same directory as the mojom tool binary or
it cannot be opened, the version check is bypassed.

R=rudominer@chromium.org, rudominer

Review URL: https://codereview.chromium.org/1806173002 .