Add additional GN dependencies in the "mojom" target. (take2)
The Mojom generator needs to be re-run if the frontend of the compiler
changes.

Also, if generate_type_info=true then we want to make sure that
mojom_types.mojom has been built.

This is intended to replace https://codereview.chromium.org/1804833002/.
The changes relative to that CL are:
- Use deps instead of inputs.

R=azani@chromium.org

Review URL: https://codereview.chromium.org/1810663003 .