Mozart: Make child view scene information invariant.
This patch is one step towards redesigning the view layout protocol.

Previously, layout served double-duty both as a means to set layout
constraints and as a means to get the view's current SceneToken
(formerly its Surface).  Now we remove the latter responsibility
by having the View Manager create a "stub scene" which wraps the view's
own scene from the perspective of its container.

This change makes it possible for the view manager to prevent leakage
of a view's actual underlying SceneToken to its container and allows
it to apply clips to the child's content.  The container only ever
sees the wrapped scene whose identity is invariant as long as the child
remains attached.

Introduced a new |OnChildAttached| event to inform the container of
its child's scene.  This may be extended with other embedding information
in the future.

The next step is to make layout feed-forward and to start using scene
versions for synchronization.

BUG=
R=abarth@google.com

Review URL: https://codereview.chromium.org/1809233003 .