Fix mac build.
This disables -Wnewline-eof on mac. It has rotted a couple of times before and
I've added the newlines back. But now the generated .mojom.cc files need them
and it just seems silly to fail a working build for lack of a newline at EOF
when the linux build doesn't require it.

BUG=
R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/1808413002 .