Dart: Temporary fix for example test
Content handler shutdown hangs if the service isolate hasn't finished
starting up when the VM wants to shutdown.

It looks like we block to wait for the service isolate to finish
starting up before calling Dart_Cleanup(), but maybe that isn't
sufficient, somehow?

Disabling the observatory removes the hang, but it would be better
if we could leave the observatory enabled for these tests.

related #723

Review URL: https://codereview.chromium.org/1831263003 .