C++ Bindings: Remove 'inline' arg from generated union serialization functions.
Previous to this CL, you could serialize a union 'inline' or
'out of line'.  This is never useful (you always want to serialize
'inline', and the serialization routine will serialize 'out of line'
when it tries to serialize a union inside a union) outside of unittests,
and is never expected to be used by a user of the serialization API.

This CL basically removes this argument from GetSerializedSize_() and
SerializeUnion_(), and updates its consumers.

R=azani@chromium.org, viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/1833033003 .