Mojom frontend: Don't stop after processing the first method.
This patch fixes a bug introduced in https://codereview.chromium.org/1833593002 but never deployed. This was a logic error that caused us to quit after processing the packing data for the first method in in an interface.

In this patch we fix the bug, add a test for it, and also update the sha1 so that we are actually deploying the fix.  The changes to the .mojom.dart files must be from some earlier change that was never deployed.

R=vardhan@google.com

Review URL: https://codereview.chromium.org/1832143003 .