Mozart: Replace layout protocol with view properties.
ViewProperties offer a simple mechanism whereby a View may provide
parameters and contextual information to its children.  Currently
this consists of the child's size in pixels and the device pixel
ratio.  Later we may also include properties such as locale, theme
accent colors, or whatever else is needed to make embedding appear
relatively seamless.

This change temporarily removes the ability for child views to resize
themselves autonomously and report those changes back to their parent.
Don't worry, it'll come back as soon as we have a better idea of how
layout negotiation should take place.

This change completes the transition to using Scene version information
as a means of cross-view state synchronization.  For example, the
parent can specify a version which it would like the child to produce
alongside an updated set of view properties.  Then the parent simply
publishes a scene update to establish a dependencies on that version
of the child.  As soon as the child has produced the desired version,
the parent's rendering will be unblocked and the new content will
become visible.  (The parent may use node combinators to avoid blocking
if desired.)

Actually, this description is slightly inaccurate.  To ensure an
orderly evolution of child view states, the view manager interposes a
wrapper scene (called the "stub scene") between the parent and the child.
The parent actually sets the version of this stub scene.  Meanwhile the
view manager uses an incrementing counter to specify the child's actual
scene version.

This change includes a significant refactoring of the ViewState and
ViewTreeState classes so as to leverage the new ViewContainerState
class to reduce code duplication.

BUG=
R=abarth@google.com

Review URL: https://codereview.chromium.org/1835803002 .