Mojom parser: Fix snippet algorithm..
Line 269 of tokens.go was not being tested and it had a bug that causes a panic when attempting to compute the snippet for a token at the end of a long line.

This patch fixes the bug and adds a test for it.

The .mojom.dart files definitely should not have changed so I think there must be some unrelated bug in the stability of the runtime type info generation that I will have to investigate.

R=azani@chromium.org

Review URL: https://codereview.chromium.org/1837353002 .