Mojom frontend: Validate and compute interface versions
We compute validate and set the |min_version| field of MojomMethod and the |CurrentVersion| field of MojomInterface.

We do not yet consume the data in the old backend.

- In computed_data.go we update StructFieldMinVersionError so that it uses appropriate wording
when issuing an error messge dealing with a Method parameter instead of a regular struct field.

- In computed_data.go we added the method computeInterfaceVersion(). This validates the MinVersions
of each of the methods and it sets the version of an interface to be equal to the maximum value of the
MinVersions of each of the methods and each of their parameters and response parameters.

- In user_defined_types.go we introduced |VersionedDeclarationData| to hold the MinVersion value
so that StructField and MojomMethod could share this.

- In user_defined_types.go we added a new field |methods_in_ordinal_order| to MojomInterface because
iterating through the methods in ordinal order is necessary in computeInterfaceVersion().

- In serialization.go we serialize the curent_version field of MojomInterface and the min_version
field of MojomMethod

- We added many test cases for all of this.

- We updated the sha1s and it is expected that some of the .mojom.dart files changed since the serialization has changed.

BUG=#713
R=azani@chromium.org

Review URL: https://codereview.chromium.org/1838023002 .