Mojom compiler frontend: Populate the declaration_order field for Methods.
This is a pretty minor change. We have not been populating the declaration_order field for MojomMethod in the intermediate representation even though the infrastructure is there to do that. We rectify this here. This simply allows the generators to write methods into the generated source in original lexical order if they wish to do so.

We do not yet consume this in the old backend. (Currently the generators in the old backend are writing the methods in ordinal order. This differs from what version 1 of the compiler used to do I think.)

R=azani@chromium.org

Review URL: https://codereview.chromium.org/1844103002 .