Motown: Add network reader to factory service NOTE: services/media/framework/incident.* is properly introduced in another CL (https://codereview.chromium.org/1847703002/). Feedback on that class should happen there.
Motown is moving away from accepting URLs that specify what content to
play to accepting a reader service instead. This CL defines the
interface (SeekingReader) for such a service and implements that
interface on top of the network service.

In the long run, we'll want SeekingReader, or something like it, to be
a common interface exposed by multiple services (e.g. network and file).
For now, it's defined as a media-specific interface that's implemented
by the media factory service.

Later CLs will introduce the code that integrates the network service
into Motown.

R=kulakowski@chromium.org

Review URL: https://codereview.chromium.org/1845423002 .