Motown: Replace Event and Watched with Incident and CallbackJoiner, etc Motown introduced the 'Event' class, which was used for simplifying a number of tasks relating to asynchronous operation, including: - Responding to the completion of multiple parallel operations - Pipelined operations (waiting for one operation to complete before   starting another). - State machines (by responding to exactly one of a set of possible   events). In addition to Event, Motown employed a Watched class, which signals Events when a wrapped (Watched) value changes. Event and Watched were not well-received in code review for various reasons.
This CL deletes Event and Watched and replaces them with other stuff:

The Incident class is a little like Event, but has simpler lifetime
characteristics and a narrower scope of concerns. An Incident is
basically a publisher of state transitions. Subscribers can get a
callback when an Incident 'occurs'. This is useful for handling
pipelined operations.

The CallbackJoiner class is specifically for dealing with multiple
parallel operations. It provides a simple way to get a callback when
a set of operations are completed.

The state machine in MediaPlayerImpl, which previously used Events, now
uses an explicit state variable and an 'Update' function that takes
action on events based on state.

For now, Incident and CallbackJoiner are part of media/framework. They
can be moved to a more generic location if there's demand for them.

R=kulakowski@chromium.org

Review URL: https://codereview.chromium.org/1847703002 .