Mozart: Rename scene graph classes.
Removed the confusing *Def suffix on Node and Resource leaving it only
on SceneDef.  In an earlier design, the Def classes were intended to be
mutable representations of client-supplied content.  Later it became
clear that it was desirable to retain immutable copies of this content
instead so that they could be shared by multiple versions.

BUG=
R=abarth@google.com

Review URL: https://codereview.chromium.org/1874593002 .