EDK: Share "MockSimpleDispatcher" and use it in place of TrivialDispatcher.
We had two copies of TrivialDispatcher, and I didn't want to add a
third, and sharing TrivialDispatcher seemed kind of silly, when we have
a slightly more powerful/general MockSimpleDispatcher. (This introduces
a slightly suboptimal test dependency: the test for Dispatcher depends
on MockSimpleDispatcher, which depends on SimpleDispatcher, which
depends on Dispatcher. Oh well.)

R=azani@chromium.org

Review URL: https://codereview.chromium.org/1909483006 .