Mojom frontend: An algorithm for detecting ill-founded nodes in a graph.
We add a utility containing an algorithm for detecting ill-founded nodes in a two-sorted directed graph.

See the paper "Well-Founded Two-Sorted Directed Graphs" (https://goo.gl/ipxFKu)
for background and detailed definitions. The proofs still need to be added.

The intended use of this is to detect unavoidable cycles in the Mojom type graph for as described in
https://github.com/domokit/mojo/issues/694

BUG=#694
R=azani@chromium.org

Review URL: https://codereview.chromium.org/1908273003 .