EDK: Add Dispatcher::SupportsEntrypointClass().
But don't actually use it yet.

The problem we need to solve is: With rights, we'd like to do the rights
check in a common place (in Core) for each entrypoint. However, in the
case that the rights check fails (i.e., the handle doesn't have the
required right(s)), we prefer to say "invalid argument" -- rather than
"permission denied" -- if the dispatcher does not support that
entrypoint.

So what we do is the following, if an entrypoint |Foo()| is called:
* Get the dispatcher |d|. (If this fails, then "invalid argument".)
* Do the rights check.
* If the rights check passes:
  * Call |d->Foo()|.
  * If |d| supports |Foo()|, it does its thing (and yields whatever
    result).
  * If not, then |d->Foo()| must return "invalid argument".
* If the rights check fails, then:
  * Check |d->SupportsEntrypointClass(<class of Foo()>)|.
  * If |Foo()|'s class is supported, then "permission denied".
  * Else "invalid argument".

Note that we could of course call |d->SupportsEntrypointClass()| even
when the rights check passes. However, this is an additional call, and
penalizes the common/important (non-error) case, so we prefer to just
call |d->Foo()|.

R=azani@chromium.org

Review URL: https://codereview.chromium.org/1915153002 .