Delete the |complete_type_set| field from mojom_types.mojom
The |complete_type_set| field was deprecated and not being populated.
We also delete the code used to populate this field: The functions FindReachableTypes() in user_defined_types.go and ReferencedUserDefinedTypes() from the TypeRef interface in types.go.

We also delete the struct ServiceTypeInfo since it would now contain only a single string field and we replace
map<string, ServiceTypeInfo> services_by_name with map<string, string> services, moving the string field
from ServiceTypeInfo directly into the value of the map.

This work is in preparation for adding new code to analyze the type graph for the purpose of finding cycles. The code we are deleting also analyzed the type graph but in a different way. Since the older way of analyzing the type graph is now obsolete, it will reduce confusion to remove it before adding a newer way to analyze the type graph.

bug=#694

R=azani@chromium.org

Review URL: https://codereview.chromium.org/1916863003 .