ConnectToApplicationDeprecated() conversion/removal, part 1.
In these cases, they just connect to an application and don't do
anything with the connection (so the net result is really to start the
application, and keep it alive since ApplicationConnection objects are
owned/kept alive by the ApplicationImpl).

So it suffices to use the Shell's ConnectToApplication(), get a
ServiceProvider for the application, and keep that alive.

R=vardhan@google.com

Review URL: https://codereview.chromium.org/1914053005 .