ConnectToApplicationDeprecated() conversion/removal, part 3.
The last use of ConnectToApplicationDeprecated() is complicated, since
both ServiceProvider directions are used.

We hang on to ServiceProviderImpls (which are bound to requests whose
corresponding handles were sent over ConnectToApplication()) and keep
them alive (and of course service them), so that the TestB that we
connected to can make requests on us.

Instead of passing the ApplicationConnect* to the TestAImpl, we just
pass it the ServiceProvider handle ("pointer"), so that the TestAImpl
can connect to the TestB services.

R=vardhan@google.com

Review URL: https://codereview.chromium.org/1924673002 .