Make it possible to write a message pipe endpoint's peer into it.
This is a nonsensical thing to do, but there's no real reason to
disallow it. (Similar nonsense can be achieve anyway.)

This requires a bit of unfortunate plumbing, but maybe it's less
unfortunate than the existing plumbing. (I also need to replace
"DispatcherTransport" with "HandleTransport", which will carry handle
rights.)

R=azani@chromium.org
BUG=fixes #767

Review URL: https://codereview.chromium.org/1943123002 .